Task(#19): Move api segments from AnimeApi.kt to a new file
