+++ app/src/main/java/com/jsmirabal/animeinfo/data/service/AnimeService.kt
@@ -1,6 +1,6 @@
 package com.jsmirabal.animeinfo.data.service
 
-import com.jsmirabal.animeinfo.data.service.api.AnimeApi
+import com.jsmirabal.animeinfo.data.service.api.Top
 import com.jsmirabal.animeinfo.data.service.model.DataLayerError
 import com.jsmirabal.animeinfo.data.service.model.DataTopItems
 import com.jsmirabal.animeinfo.domain.core.ResultWrapper
@@ -8,9 +8,9 @@ import com.jsmirabal.animeinfo.domain.model.AnimeDetail
 
 interface AnimeService {
     suspend fun fetchTopItems(
-        type: AnimeApi.Type,
-        subType: AnimeApi.SubType = AnimeApi.SubType.NO_SUB_TYPE, // Optional
-        page: String = "1"
+        type: Top.Type,
+        subType: Top.SubType = Top.SubType.NO_SUB_TYPE, // Optional
+        page: String = "1" // Optional
     ): ResultWrapper<DataTopItems, DataLayerError>
 
     suspend fun fetchAnimeDetail(id: String): ResultWrapper<AnimeDetail, DataLayerError>
