+++ app/src/main/java/com/jsmirabal/animeinfo/data/service/AnimeServiceImpl.kt
@@ -1,15 +1,14 @@
 package com.jsmirabal.animeinfo.data.service
 
 import com.jsmirabal.animeinfo.data.service.api.AnimeApi
-import com.jsmirabal.animeinfo.data.service.api.AnimeApi.SubType
-import com.jsmirabal.animeinfo.data.service.api.AnimeApi.Type
+import com.jsmirabal.animeinfo.data.service.api.Top
 import com.jsmirabal.animeinfo.data.service.model.DataLayerError
 import com.jsmirabal.animeinfo.domain.core.ResultWrapper
 import retrofit2.HttpException
 
 class AnimeServiceImpl(private val animeApi: AnimeApi) : AnimeService {
 
-    override suspend fun fetchTopItems(type: Type, subType: SubType, page: String) = try {
+    override suspend fun fetchTopItems(type: Top.Type, subType: Top.SubType, page: String) = try {
         ResultWrapper.Success(animeApi.fetchTopItems(type.get(), subType.get(), page))
     } catch (e: Exception) {
         getError(e)
