+++ app/src/main/java/com/jsmirabal/animeinfo/domain/model/AnimeDelegates.kt
